requirements.txt
django
django-rest-framework
django-extensions
celery
django-celery-results
psycopg2-binary
redis
